Update folder structure
폴더 구조 변경
Atomic Desgin을 일부 컴포넌트에 반영
전체 컴포넌트에 대해서는 팀원 전체가
디자인을 보고 반영한다.
