+++ client/src/components/Molecules/Carousel/RemoveButton/index.jsx
