+++ client/src/pages/Register/template/InsertInfo/index.jsx
@@ -3,12 +3,12 @@ import styled from 'styled-components'
 
 import PageBase from '../../../../components/PageBase'
 import Button from '../../../../components/BoxButton'
-import Carousel from '../../../../components/Carousel'
+import Carousel from '../../../../components/Molecules/Carousel'
 import InputBox from '../../../../components/InputBox'
-import MoneyBox from '../../../../components/MoneyBox'
+import MoneyBox from '../../../../components/Molecules/MoneyBox'
 import ItemDescription from '../../../../components/ItemDescription'
 import TermSelector from '../../../../components/RegisterTermSelector'
-import ToggleButton from '../../../../components/ToggleButton'
+import ToggleButton from '../../../../components/Atoms/ToggleButton'
 
 import { termList, itemDescription } from '../../constants'
 import { idxNotSelected, isArrayEmpty, strEmpty } from '../../../../utils/validator.js'
